Randomize bounce ray direction based on the Lambertian diffuse model
This follows how perfectly diffuse materials reflect light. Intuitively, this
corresponds to NdotL lighting where the surface is shaded brighter when the
surface normal and the light are facing closer to each other.

In the current scene, this darkens the areas under the smaller spheres because
the light is more likely to bounce from the ground back into the smaller
spheres, due to the weighting in favor of the hit surface's normal, instead of
terminating into the "sky". This means more light is absorbed by the surfaces,
in accordance with the current "material".